Trader/requirements.txt
PyYAML~=6.0.2
websockets~=14.1
requests~=2.32.3
pandas~=2.2.3